Create dist.py
Script to convert RSSI to feet. The following equation is used.
RSSI (dBm) = -10n log10(d) + A, 
d: distance in metres 
A: received signal strength in dBm at 1 meter
n: propagation constant or path-loss exponent (Free space has n =2 for reference).
An equation is also included for the RSSI data that was manually collected. It's commented out because the results are not as accurate as the new equation.